+++ src/main/java/it/model/Campo.java
@@ -6,6 +6,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -28,7 +29,7 @@ public class Campo {
     private String tipologia;
 
     @Column(name="prezzo")
-    private Double prezzo;
+    private BigDecimal prezzo;
 
     @Column(name="coperto")
     private Boolean coperto;
